% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemSymWords.R
\name{chemWordPNG}
\alias{chemWordPNG}
\title{Create a PNG representation using chemical elements}
\usage{
chemWordPNG(w, sym = elements, line1 = c("Name", "center"),
  line2 = c("Atomic_Number", "center"), line3 = c("Atomic_Mass", "center"),
  additional = NULL, colorInvert = FALSE, rounded = FALSE,
  f = "chemWord.png")
}
\arguments{
\item{w}{Character.  A word to convert to chemical elements}

\item{sym}{Data Frame.  A data frame containing chemical information.  Defaults to elements}

\item{line1}{Vector containing two elements.  Top row label description.  First element
corresponds to a column ID in the \code{sym} data frame.  Second element corresponds
to the horizontal positioning (left, right, center).  To remove this line from the PNG
output set it to c(NA, NA)}

\item{line2}{Vector containing two elements.  Second row label description.  See \code{line1} description}

\item{line3}{Vector containing two elements.  Bottom row label description.  See \code{line1} description}

\item{additional}{Data Frame.  Optional additional element table.  Data frame which must
have a \code{Symbol} column containing the chemical symbol and can also contain
\code{Atomic_Number}, \code{Name} and \code{Atomic_Mass} for inclusion in the PNG}

\item{colorInvert}{Boolean.  Invert colors.  If true then elements will be white colored on a
black background.  Default is false (black text on white background)}

\item{rounded}{Boolean.  Rounded box corners.  Default is false}

\item{f}{Character.  Filename for output file (including path)}
}
\value{
Boolean.  True if file generated, false if word could not be represented by chemical elements
}
\description{
Output a PNG file of chemical elements for a particular word
}

