% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemSymWords.R
\name{chemWordRecurse}
\alias{chemWordRecurse}
\title{recurse a string identifying chemical elements}
\usage{
chemWordRecurse(w, t, sym = elements)
}
\arguments{
\item{w}{Character.  Text to search for chemical elements}

\item{t}{List.  Running list of partial matches for current search path}

\item{sym}{Data frame.  A data frame containing chemical information}
}
\value{
List of results.  Each list item refers to a positive result in representing the
  input string as a series of elements.  List items are vectors corresponding to row
  positions of the elements in \code{sym}.
}
\description{
Recursive function to find chemical elements within a string
}

