% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elements}
\alias{elements}
\title{Chemical Elements}
\format{A data frame with 118 rows and 4 variables:
\describe{
  \item{Atomic_Number}{atomic number}
  \item{Symbol}{chemical symbol}
  \item{Name}{chemical name}
  \item{Atomic_Mass}{atomic mass}
}}
\usage{
elements
}
\description{
A dataframe containing details for the first 118 elements
}
\keyword{datasets}

