% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemSymWords.R
\name{chemWord}
\alias{chemWord}
\title{Represent a string using chemical elements}
\usage{
chemWord(w, output = "min", sym = elements)
}
\arguments{
\item{w}{Character.  A word to convert to chemical elements}

\item{output}{Character.  Specify the output.  \code{min} returns the fewest elements,
\code{max} returns the most elements, \code{list} returns a list of all hits}

\item{sym}{Data Frame.  A data frame containing chemical information.  Defaults to elements}
}
\value{
Character vector of elements if successful or FALSE if not all characters can be assinged to elements
}
\description{
Represent a string using chemical elements
}

