% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemSymWords.R
\name{chemWordPNG}
\alias{chemWordPNG}
\title{Create a PNG representation using chemical elements}
\usage{
chemWordPNG(w, sym = elements, f = "chemWord.png")
}
\arguments{
\item{w}{Character.  A word to convert to chemical elements}

\item{sym}{Data Frame.  A data frame containing chemical information.  Defaults to elements}

\item{f}{Character.  Filename for output file (including path)}
}
\value{
Boolean.  True if file generated, false if word could not be represented by chemical elements
}
\description{
Output a PNG file of chemical elements for a particular word
}

