% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemSymWords.R
\name{chemWordPNG}
\alias{chemWordPNG}
\title{Create a PNG representation using chemical elements}
\usage{
chemWordPNG(w, sym = elements, additional = NULL, f = "chemWord.png")
}
\arguments{
\item{w}{Character.  A word to convert to chemical elements}

\item{sym}{Data Frame.  A data frame containing chemical information.  Defaults to elements}

\item{additional}{Data Frame.  Optional additional element table.  Data frame which must
have a \code{Symbol} column containing the chemical symbol and can also contain
\code{Atomic_Number}, \code{Name} and \code{Atomic_Mass} for inclusion in the PNG}

\item{f}{Character.  Filename for output file (including path)}
}
\value{
Boolean.  True if file generated, false if word could not be represented by chemical elements
}
\description{
Output a PNG file of chemical elements for a particular word
}

